refactor password-reset + schedules backend
