+++ backend/routes/password-reset/helpers.js
@@ -1,0 +1,6 @@
+/* eslint-disable */
+const crypto = require('crypto');
+
+const generateResetToken = () => crypto.randomBytes(32).toString('hex');
+
+module.exports = { generateResetToken };
