+++ backend/routes/schedules/create.js
@@ -1,0 +1,76 @@
+/* eslint-disable */
+const express = require('express');
+const router = express.Router();
+const pool = require('../../config/database');
+const authenticateToken = require('../../middleware/auth');
+const { validateSchedule, canEditChildSchedule } = require('./validation');
+
+// Create a new schedule entry
+router.post('/', authenticateToken, async (req, res) => {
+  const client = await pool.connect();
+  try {
+    const { child_id, class_id, date, start_time, duration_hours, activity, notes } = req.body;
+
+    const validationErrors = validateSchedule(req.body);
+    if (validationErrors.length > 0) {
+      return res.status(400).json({ errors: validationErrors });
+    }
+
+    const canEdit = await canEditChildSchedule(req.user.id, req.user.role, child_id);
+    if (!canEdit) {
+      return res
+        .status(403)
+        .json({ error: "You do not have permission to edit this child's schedule" });
+    }
+
+    await client.query('BEGIN');
+
+    const classChildResult = await client.query(
+      'SELECT 1 FROM class_children WHERE child_id = $1 AND class_id = $2',
+      [child_id, class_id]
+    );
+
+    if (classChildResult.rows.length === 0) {
+      await client.query('ROLLBACK');
+      return res.status(400).json({ error: 'Child is not assigned to this class' });
+    }
+
+    const conflictResult = await client.query(
+      `
+      SELECT id FROM schedules 
+      WHERE child_id = $1 AND date = $2 
+      AND (
+        (start_time <= $3 AND start_time + (duration_hours || ' hours')::interval > $3) OR
+        (start_time < ($3 + ($4 || ' hours')::interval) AND start_time + (duration_hours || ' hours')::interval >= ($3 + ($4 || ' hours')::interval)) OR
+        (start_time >= $3 AND start_time + (duration_hours || ' hours')::interval <= ($3 + ($4 || ' hours')::interval))
+      )
+    `,
+      [child_id, date, start_time, duration_hours]
+    );
+
+    if (conflictResult.rows.length > 0) {
+      await client.query('ROLLBACK');
+      return res.status(409).json({ error: 'Time conflict with existing schedule entry' });
+    }
+
+    const result = await client.query(
+      `
+      INSERT INTO schedules (child_id, class_id, date, start_time, duration_hours, activity, notes, created_by, updated_by)
+      VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $8)
+      RETURNING *
+    `,
+      [child_id, class_id, date, start_time, duration_hours, activity, notes, req.user.id]
+    );
+
+    await client.query('COMMIT');
+    res.status(201).json(result.rows[0]);
+  } catch (err) {
+    await client.query('ROLLBACK');
+    console.error('Error creating schedule:', err);
+    res.status(500).json({ error: 'Failed to create schedule entry' });
+  } finally {
+    client.release();
+  }
+});
+
+module.exports = router;
